refactor: reviewed changes
