+++ frontend/lib/data/repositories/atproto/atproto_repository_local.dart
@@ -1,10 +1,10 @@
 import 'package:bluesky/bluesky.dart' as bsky;
 import 'package:atproto/atproto_oauth.dart';
 import 'package:frontend/data/services/local/local_data_service.dart';
-import 'package:frontend/shared/exceptions/not_authorized_exception.dart';
+import 'package:frontend/shared/exceptions/unauthorized_exception.dart';
 import 'package:frontend/shared/models/post/post.dart';
 import 'package:frontend/domain/models/post/post.dart' as domain_models;
-import 'package:frontend/shared/util/convert_feed_to_domain_posts.dart';
+import 'package:frontend/internal/feed/feed.dart';
 import 'package:frontend/utils/result.dart';
 import './atproto_repository.dart';
 import 'package:atproto/atproto.dart' as atp;
@@ -88,26 +88,24 @@ class AtProtoRepositoryLocal extends AtProtoRepository {
 
   @override
   Future<Result<void>> createPost(Post post) async {
-    if (authorized) {
-      return await _localDataService.createPost(atProto!, post);
-    } else {
-      return Result.error(NotAuthorizedException('createPost'));
+    if (!authorized) {
+      return Result.error(UnauthorizedException('createPost'));
     }
+    return await _localDataService.createPost(atProto!, post);
   }
 
   @override
   Future<Result<List<domain_models.Post>>> getFeed() async {
-    if (authorized) {
-      final bskyFeed = _localDataService.getFeed();
+    if (!authorized) {
+      return Result.error(UnauthorizedException('getFeed'));
+    }
+    final bskyFeed = _localDataService.getFeed();
 
-      switch (bskyFeed) {
-        case Ok<bsky.Feed>():
-          return Result.ok(convertFeedToDomainPosts(bskyFeed.value));
-        case Error<bsky.Feed>():
-          return Result.error(bskyFeed.error);
-      }
-    } else {
-      return Result.error(NotAuthorizedException('getFeed'));
+    switch (bskyFeed) {
+      case Ok<bsky.Feed>():
+        return Result.ok(convertFeedToDomainPosts(bskyFeed.value));
+      case Error<bsky.Feed>():
+        return Result.error(bskyFeed.error);
     }
   }
 }
