+++ frontend/lib/data/repositories/atproto/atproto_repository_remote.dart
@@ -1,8 +1,8 @@
 import 'package:bluesky/bluesky.dart' as bsky;
-import 'package:frontend/shared/exceptions/not_authorized_exception.dart';
+import 'package:frontend/shared/exceptions/unauthorized_exception.dart';
 import 'package:frontend/shared/models/post/post.dart';
 import 'package:frontend/domain/models/post/post.dart' as domain_models;
-import 'package:frontend/shared/util/convert_feed_to_domain_posts.dart';
+import 'package:frontend/internal/feed/feed.dart';
 import 'package:frontend/utils/result.dart';
 import 'package:frontend/data/services/api/api_client.dart';
 import './atproto_repository.dart';
@@ -74,26 +74,24 @@ class AtProtoRepositoryRemote extends AtProtoRepository {
 
   @override
   Future<Result<void>> createPost(Post post) async {
-    if (authorized) {
-      return await _apiClient.createPost(atProto!, post);
-    } else {
-      return Result.error(NotAuthorizedException('createPost'));
+    if (!authorized) {
+      return Result.error(UnauthorizedException('createPost'));
     }
+    return await _apiClient.createPost(atProto!, post);
   }
 
   @override
   Future<Result<List<domain_models.Post>>> getFeed() async {
-    if (authorized) {
-      final bskyFeed = await _apiClient.getFeed(atProto!.oAuthSession!);
+    if (!authorized) {
+      return Result.error(UnauthorizedException('getFeed'));
+    }
+    final bskyFeed = await _apiClient.getFeed(atProto!.oAuthSession!);
 
-      switch (bskyFeed) {
-        case Ok<bsky.Feed>():
-          return Result.ok(convertFeedToDomainPosts(bskyFeed.value));
-        case Error<bsky.Feed>():
-          return Result.error(bskyFeed.error);
-      }
-    } else {
-      return Result.error(NotAuthorizedException('getFeed'));
+    switch (bskyFeed) {
+      case Ok<bsky.Feed>():
+        return Result.ok(convertFeedToDomainPosts(bskyFeed.value));
+      case Error<bsky.Feed>():
+        return Result.error(bskyFeed.error);
     }
   }
 }
