+++ frontend/lib/main_staging.dart
@@ -9,7 +9,7 @@ import 'main.dart';
 /// Staging config entry point.
 /// Launch with `flutter run --target lib/main_staging.dart`.
 /// Uses remote data from a server.
-void main() async {
+void main() {
   Logger.level = Level.all;
 
   usePathUrlStrategy();
