+++ frontend/lib/shared/exceptions/missing_env.dart
@@ -1,0 +1,8 @@
+class MissingEnvException implements Exception {
+  final String env;
+  MissingEnvException(this.env);
+
+  @override
+  String toString() =>
+      'MissingEnvException: Missing environment variable - $env';
+}
