+++ tourbillon/src/algos/boshi.ts
@@ -6,6 +6,13 @@ import { Agent } from '@atproto/api'
 // max 15 chars
 export const shortname = 'boshi'
 
+/**
+ * https://docs.bsky.app/docs/api/app-bsky-feed-get-feed-skeleton
+ * Handler for server.app.bsky.feed.getFeedSkeleton method
+ * @constructor
+ * @param {string} ctx - Context
+ * @param {string} params - Params for a Get Feed Skeleton Query request
+ */
 export const handler = async (
   ctx: AppContext,
   params: FeedSkeletonQueryParams,
@@ -31,6 +38,14 @@ export const handler = async (
   }
 }
 
+/**
+ * https://docs.bsky.app/docs/api/app-bsky-feed-get-feed
+ * Handler for server.app.bsky.feed.getFeed method
+ * Also queries for the user profiles of the posts
+ * @constructor
+ * @param {string} ctx - Context
+ * @param {string} params - Params for a Get Feed Query request
+ */
 export const feedHandler = async (ctx: AppContext, params: FeedQueryParams) => {
   let builder = ctx.db
     .selectFrom('post')
